# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sample Ecommerce App
  description: >
    'This is a ***sample ecommerce app API***.  You can find out more about Swagger at [swagger.io](http://swagger.io).
    Description supports markdown markup. For example, you can use the `inline code` using back ticks.'
  termsOfService: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  contact:
    email: support@packtpub.com
  license:
    name: MIT
    url: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: /api/v1  # Relative path which maps to localhost:8080 on dev
    description: Current server
...
paths:
...
  /customers/{id}/orders:
    get:
      tags:
        - order
      summary: Retrieves the orders for a specific customer
      description: Retrieves a collection of orders associated with the given customer ID.
      operationId: getCustomerOrders
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch. Returns a collection of orders.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
            application/json:
              schema:
                type: array # Ensure the JSON schema is an array, matching the XML type
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags:
        - order
      summary: Creates a new order for a customer
      description: Creates a new order for a customer with the given order request.
      operationId: addOrder
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: true
        description: New Order Request object
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/OrderReq'
          application/json:
            schema:
              $ref: '#/components/schemas/OrderReq'
      responses:
        201:
          description: Order added successfully
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        406:
          description: If payment is not authorized.
          content: { }
...
  /orders:
    get:
      tags:
        - admin
      summary: Retrieves all orders in the system
      description: Retrieves all orders. Typically an admin-only endpoint.
      operationId: getAllOrders
      responses:
        200:
          description: If order exists.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        404:
          description: Order doesn't exist for given customer.
          content: { }
  /orders/{id}:
    get:
      tags:
        - order
      summary: Retrieves the order of given order ID
      description: Retrieves orders of given order ID
      operationId: getByOrderId
      parameters:
        - name: id
          in: path
          required: true
          description: Order Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: If order exists.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          description: Order doesn't exist for given customer.
          content: {}
...
components:
  schemas:
    UUID:
      type: string
      description: A standard unique identifier
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      example: 00000000-0000-0000-0000-000000000001
    BigDecimal:
      type: string
      description: High-precision decimal number. Transmitted as a string to preserve precision.
      example: "123.45"
      # Pattern for validation, ensures the string contains only a number, an optional minus sign at the beginning, and an optional decimal part:
      pattern: '^-?\d+(\.\d+)?$'
...
    CustomerBase:
      type: object
      description: Base properties of a customer object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
      xml:
        name: CustomerBase
...
    Customer:
      type: object
      description: "Represents a complete customer resource, including its ID"
      allOf:
        - $ref: '#/components/schemas/CustomerBase'
      required:
        - id
      properties:
        id:
          description: The id of the customer
          $ref: '#/components/schemas/UUID'
      xml:
        name: Customer
    AddressBase:
      type: object
      description: Base properties for an address
      required:
        - street
        - city
        - zipcode
      properties:
        number:
          type: string
          description: house or flat number
          maxLength: 20
        residency:
          type: string
          description: Society or building name (optional)
        street:
          type: string
          description: street name
          minLength: 1
          maxLength: 80
        city:
          type: string
          description: city name
          minLength: 1
          maxLength: 40
        state:
          type: string
          description: state name
          maxLength: 40
        country:
          type: string
          description: country name
          maxLength: 40
        zipcode:
          type: string
          description: postal code
          minLength: 1
          maxLength: 20
      xml:
        name: AddressBase
...
    Address:
      type: object
      description: "Represents a complete address resource, including its ID and associated Customer ID."
      allOf:
        - $ref: '#/components/schemas/AddressBase'
      required:
        - id
        - customerId
      properties:
        id:
          description: Address identifier
          $ref: '#/components/schemas/UUID'
        customerId:
          description: The customer ID associated with this address.
          $ref: '#/components/schemas/UUID'
      xml:
        name: Address
    CardDetailsBase:
      type: object
      description: Base properties for a credit/debit card.
      required:
        - cardNumber
        - expires
        - cvv
      properties:
        cardNumber:
          type: string
          description: Card Number (The full number should only be sent in the request, not returned)
        expires:
          type: string
          description: Expiration date (e.g., MM/YY or MM/YYYY)
          example: 01/29
          pattern: '^(0[1-9]|1[0-2])\/\d{2}(\d{2})?$' # Example: MM/YY or MM/YYYY
        cvv:
          type: string
          description: CVV code (Only sent in the request, never stored or returned)
          example: 123
          pattern: '^\d{3,4}$' # 3 or 4 digits
          writeOnly: true
      xml:
        name: CardDetailsBase
...
    Card:
      type: object
      description: Represents a registered Card resource, with masked details.
      allOf:
          - $ref: '#/components/schemas/CardDetailsBase'
      required:
        - id
        - customerId
        - cardNumber
        - expires
      properties:
        id:
          description: Card identifier
          $ref: '#/components/schemas/UUID'
        customerId:
          description: Id of customer owning the card
          $ref: '#/components/schemas/UUID'
        cardNumber:
          type: string
          description: Card Number (Masked for security)
          example: '************1234'
        expires:
          type: string
          description: Expiration date (MM/YY or MM/YYYY)
        cvv:
          type: string
          description: CVV code (Masked response for security)
          readOnly: true
          example: '***'
      xml:
        name: Card
...
    Item:
      type: object
      description: Item in shopping cart
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          description: Product Identifier for this item
          $ref: '#/components/schemas/UUID'
        quantity:
          type: integer
          description: The item quantity
          format: int32
          minimum: 0
        unitPrice:
          description: The item's price per unit
          allOf:
            - $ref: '#/components/schemas/BigDecimal'
      xml:
        name: Item
...
    OrderBase:
      type: object
      description: Base properties for an order, including the items list.
      properties:
        items:
          type: array
          description: Collection of items for the order.
          items:
            $ref: '#/components/schemas/Item'
      xml:
        name: OrderBase
    OrderReq:
      type: object
      description: Request object for a new order
      allOf:
        - $ref: '#/components/schemas/OrderBase'
      required:
        - addressId
        - cardId
      properties:
        addressId:
          description: Identifier of the customer's shipping address
          $ref: '#/components/schemas/UUID'
        cardId:
          description: Identifier of the customer's payment card
          $ref: '#/components/schemas/UUID'
      xml:
        name: OrderReq
    Order:
      type: object
      description: Represents an order
      allOf:
        - $ref: '#/components/schemas/OrderBase'
      required:
        - id
        - date
        - total
        - status
        - payment
        - shipment
      properties:
        id:
          description: Order identifier
          $ref: '#/components/schemas/UUID'
        customer:
          $ref: '#/components/schemas/Customer'
        address:
          $ref: '#/components/schemas/Address'
        card:
          $ref: '#/components/schemas/Card'
        date:
          type: string
          description: Order's data and time details
          format: date-time
        total:
          description: Order total
          allOf:
            - $ref: '#/components/schemas/BigDecimal'
        payment:
          $ref: '#/components/schemas/Payment'
        shipment:
          $ref: '#/components/schemas/Shipment'
        status:
          type: string
          description: Order Status
          enum:
            - CREATED
            - PAID
            - SHIPPED
            - PAYMENT_FAILED
            - SHIPMENT_FAILED
            - COMPLETED
      xml:
        name: Order
...
    Payment:
      type: object
      description: Object corresponding to a payment.
      properties:
        id:
          description: Payment identifier
          $ref: '#/components/schemas/UUID'
        orderId:
          description: Order Identifier belonging to this payment request
          $ref: '#/components/schemas/UUID'
        authorized:
          type: boolean
          description: Flag that specifies whether payment is authorized or not
        message:
          type: string
          description: Approval or rejection message
      xml:
        name: Payment
...
    Shipment:
      type: object
      description: Object corresponding to a shipment.
      properties:
        id:
          description: Shipment Identifier
          $ref: '#/components/schemas/UUID'
        orderId:
          description: Order Identifier
          $ref: '#/components/schemas/UUID'
        carrier:
          type: string
          description: Shipping Carrier
        trackingNumber:
          type: string
          description: Shipping Tracking Number
        estDeliveryDate:
          type: string
          description: Estimated Delivery Date
          format: date
      xml:
        name: Shipment